refactoring class airport
