+++ Java/src/test/java/AirportTest.java
@@ -7,6 +7,7 @@ import org.testng.annotations.Test;
 import Planes.MilitaryPlane;
 import Planes.PassengerPlane;
 import Planes.Plane;
+import airport.Airport;
 
 import java.util.Arrays;
 import java.util.List;
